feat(settings): UI fix and functionality fix

+++ apps/settings/lib/src/features/display/data/display_repository.dart
@@ -2,7 +2,7 @@ import 'package:mechanix_settings/src/features/display/models/types.dart';
 
 abstract class DisplayRepository {
   Future<DBusDefaultSettings> getDisplaySettings();
-  Future<void> setBrightness(int brightness);
+  Future<void> setBrightness(double brightness);
   Future<void> setAutoBrightness(bool autoBrightness);
   Future<void> setScreenTimeout(int timeout);
   Future<void> setLockScreenTimeout(int timeout);
